Hooked up all existing systems
Rishav did half the work poggies

+++ Assets/Scripts/Buildings/SourceBuilding.cs
@@ -3,8 +3,8 @@
 [System.Serializable]
 public class SourceBuilding : BuildingTemplate
 {
-    [SerializeField] private float produceValue;
-    public float ProduceValue => produceValue;
+    [SerializeField] private int produceValue;
+    public int ProduceValue => produceValue;
 }
